requirements.txt
django
django-crispy-forms
django-extensions
graphene_django
fpdf2
gunicorn
